requirements.txt
flask
Pillow
matplotlib
mxnet
gluoncv
opencv-python-headless
numpy==1.19.5